move bun logic.
